Levels/Level 2/06-image-organizer/requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
click>=8.1.0
